refactor(lab-n64g): extract LabMcpServer to mcp/server.rs
Move LabMcpServer struct, ServerHandler impl, action_schema(),
elicit_confirm(), ElicitResult, extract_error_info(), static_kind()
from cli/serve.rs to mcp/server.rs. Moves static_kind_round_trips
test too. serve.rs keeps transport/port resolution and CLI args only.
