+++ crates/lab/src/cli/serve.rs
@@ -5,20 +5,12 @@ use std::sync::Arc;
 
 use anyhow::{Context, Result};
 use clap::{Args, ValueEnum};
-use rmcp::model::{
-    CallToolRequestParams, CallToolResult, Content, CreateElicitationRequestParams,
-    ElicitationAction, ElicitationSchema, ListToolsResult, PaginatedRequestParams, PrimitiveSchema,
-    ServerCapabilities, ServerInfo, Tool,
-};
-use rmcp::service::RequestContext;
-use rmcp::{ErrorData, RoleServer, ServerHandler, ServiceExt};
-use serde_json::Value;
+use rmcp::ServiceExt;
 
 use crate::api::AppState;
 use crate::config::LabConfig;
-use crate::mcp::envelope::{build_error, build_error_extra, build_success};
-use crate::mcp::error::DispatchError;
 use crate::mcp::registry::{ToolRegistry, build_default_registry};
+use crate::mcp::server::LabMcpServer;
 
 /// Transport choices for `lab serve`.
 #[derive(Debug, Clone, Copy, ValueEnum)]
@@ -148,241 +140,10 @@ fn filter_registry(registry: ToolRegistry, services: &[String]) -> Result<ToolRe
     Ok(out)
 }
 
-/// JSON Schema for every service tool's input: `action` (required) + `params` (optional object).
-#[allow(clippy::expect_used)]
-fn action_schema() -> serde_json::Map<String, Value> {
-    serde_json::json!({
-        "type": "object",
-        "properties": {
-            "action": {
-                "type": "string",
-                "description": "Action to perform (e.g. \"movie.search\"). Use \"help\" to list all actions."
-            },
-            "params": {
-                "type": "object",
-                "description": "Action-specific parameters (varies per action)"
-            }
-        },
-        "required": ["action"]
-    })
-    .as_object()
-    .cloned()
-    .expect("schema literal is always an object")
-}
-
-/// MCP server handler — one tool per registered service.
-struct LabMcpServer {
-    registry: Arc<ToolRegistry>,
-}
-
-impl ServerHandler for LabMcpServer {
-    fn get_info(&self) -> ServerInfo {
-        ServerInfo::new(ServerCapabilities::builder().enable_tools().build())
-    }
-
-    async fn list_tools(
-        &self,
-        _request: Option<PaginatedRequestParams>,
-        _context: RequestContext<RoleServer>,
-    ) -> Result<ListToolsResult, ErrorData> {
-        let schema = Arc::new(action_schema());
-        let tools: Vec<Tool> = self
-            .registry
-            .services()
-            .iter()
-            .map(|svc| Tool::new(svc.name, svc.description, Arc::clone(&schema)))
-            .collect();
-        Ok(ListToolsResult::with_all_items(tools))
-    }
-
-    async fn call_tool(
-        &self,
-        request: CallToolRequestParams,
-        context: RequestContext<RoleServer>,
-    ) -> Result<CallToolResult, ErrorData> {
-        let service = request.name.as_ref().to_string();
-        let args = request.arguments.unwrap_or_default();
-        let action = args
-            .get("action")
-            .and_then(|v| v.as_str())
-            .unwrap_or("")
-            .to_string();
-        let params = args.get("params").cloned().unwrap_or(Value::Null);
-
-        let svc = self.registry.services().iter().find(|s| s.name == service);
-
-        // Elicitation gate: if the action is destructive and the client supports
-        // elicitation, ask for confirmation before dispatching.
-        if let Some(entry) = svc {
-            let is_destructive = entry
-                .actions
-                .iter()
-                .any(|a| a.name == action && a.destructive);
-            if is_destructive {
-                match elicit_confirm(&context, &service, &action).await {
-                    ElicitResult::Confirmed => {}
-                    ElicitResult::Declined | ElicitResult::Cancelled => {
-                        let envelope = build_error(
-                            &service,
-                            &action,
-                            "confirmation_required",
-                            &format!("action `{action}` is destructive — confirm to proceed"),
-                        );
-                        return Ok(CallToolResult::error(vec![Content::text(
-                            envelope.to_string(),
-                        )]));
-                    }
-                    ElicitResult::NotSupported => {
-                        // Client does not support elicitation — allow params["confirm"] == true
-                        // as a machine-to-machine bypass (mirrors HTTP's handle_action()).
-                        // Automated agents (CI, Claude Desktop non-interactive, headless MCP)
-                        // can opt in by passing confirm:true in the action params.
-                        if params.get("confirm").and_then(Value::as_bool) != Some(true) {
-                            let envelope = build_error(
-                                &service,
-                                &action,
-                                "confirmation_required",
-                                &format!(
-                                    "action `{action}` is destructive — pass \
-                                     {{\"confirm\":true}} in params or use a client \
-                                     that supports MCP elicitation"
-                                ),
-                            );
-                            return Ok(CallToolResult::error(vec![Content::text(
-                                envelope.to_string(),
-                            )]));
-                        }
-                    }
-                }
-            }
-        }
-
-        let start = std::time::Instant::now();
-        let result = match svc {
-            Some(entry) => (entry.dispatch)(action.clone(), params)
-                .await
-                .map_err(|te| anyhow::Error::from(DispatchError::from(te))),
-            None => Err(anyhow::anyhow!(
-                "service `{service}` has no dispatcher wired"
-            )),
-        };
-        let elapsed_ms = start.elapsed().as_millis();
-
-        match result {
-            Ok(v) => {
-                tracing::info!(surface = "mcp", service, action, elapsed_ms, "dispatch ok");
-                let envelope = build_success(&service, &action, &v);
-                Ok(CallToolResult::success(vec![Content::text(
-                    envelope.to_string(),
-                )]))
-            }
-            Err(e) => {
-                let (kind, message, extra) = extract_error_info(&e);
-                let is_fatal = matches!(kind, "internal_error" | "server_error" | "decode_error");
-                if is_fatal {
-                    tracing::error!(
-                        surface = "mcp",
-                        service,
-                        action,
-                        elapsed_ms,
-                        kind,
-                        "dispatch error"
-                    );
-                } else {
-                    tracing::warn!(
-                        surface = "mcp",
-                        service,
-                        action,
-                        elapsed_ms,
-                        kind,
-                        "dispatch error"
-                    );
-                }
-                let envelope = extra.map_or_else(
-                    || build_error(&service, &action, kind, &message),
-                    |ref extra| build_error_extra(&service, &action, kind, &message, extra),
-                );
-                Ok(CallToolResult::error(vec![Content::text(
-                    envelope.to_string(),
-                )]))
-            }
-        }
-    }
-}
-
-/// Outcome of an elicitation confirmation request.
-enum ElicitResult {
-    /// User confirmed the destructive action.
-    Confirmed,
-    /// User explicitly declined.
-    Declined,
-    /// User cancelled (closed the dialog without choosing).
-    Cancelled,
-    /// MCP client does not support the elicitation capability.
-    NotSupported,
-}
-
-/// Ask the MCP client to confirm a destructive action via elicitation.
-///
-/// Sends a form with a single required `confirm: boolean` field.
-/// Returns `NotSupported` if the client's capabilities do not include elicitation.
-async fn elicit_confirm(
-    context: &RequestContext<RoleServer>,
-    service: &str,
-    action: &str,
-) -> ElicitResult {
-    if context.peer.supported_elicitation_modes().is_empty() {
-        return ElicitResult::NotSupported;
-    }
-
-    let Ok(schema) = ElicitationSchema::builder()
-        .required_property(
-            "confirm",
-            PrimitiveSchema::Boolean(rmcp::model::BooleanSchema::default()),
-        )
-        .build()
-    else {
-        return ElicitResult::NotSupported;
-    };
-
-    let params = CreateElicitationRequestParams::FormElicitationParams {
-        meta: None,
-        message: format!(
-            "Action `{service}.{action}` is destructive and cannot be undone. \
-             Set `confirm` to true to proceed."
-        ),
-        requested_schema: schema,
-    };
-
-    match context.peer.create_elicitation(params).await {
-        Ok(result) => match result.action {
-            ElicitationAction::Accept => {
-                // Check that the user actually set confirm = true in the response.
-                let confirmed = result
-                    .content
-                    .as_ref()
-                    .and_then(|v| v.get("confirm"))
-                    .and_then(Value::as_bool)
-                    .unwrap_or(false);
-                if confirmed {
-                    ElicitResult::Confirmed
-                } else {
-                    ElicitResult::Declined
-                }
-            }
-            ElicitationAction::Decline => ElicitResult::Declined,
-            ElicitationAction::Cancel => ElicitResult::Cancelled,
-        },
-        Err(_) => ElicitResult::NotSupported,
-    }
-}
-
 async fn run_http(host: &str, port: u16, bearer_token: &str, state: AppState) -> Result<ExitCode> {
     let router =
         crate::api::router::build_router_with_bearer(state, Some(bearer_token.to_string()));
     // Parse and validate the address at bind time, not at CLI parse time.
-    // This defers host/port errors to when the HTTP transport is actually used
-    // and gives tokio a chance to surface OS-level bind errors directly.
     let addr = format!("{host}:{port}");
     let listener = tokio::net::TcpListener::bind(&addr)
         .await
@@ -403,150 +164,10 @@ async fn run_stdio(registry: Arc<ToolRegistry>) -> Result<ExitCode> {
     Ok(ExitCode::SUCCESS)
 }
 
-/// Recover a stable kind tag and message from an `anyhow::Error`.
-///
-/// Priority:
-/// 1. Downcast to [`DispatchError`] — gives structured kind + optional extras.
-/// 2. Parse `e.to_string()` as JSON `{ "kind": "…" }` — covers `ToolError`
-///    errors that were serialized to string before entering anyhow (radarr).
-/// 3. Fall back to `"internal_error"`.
-fn extract_error_info(e: &anyhow::Error) -> (&'static str, String, Option<Value>) {
-    // 1. Structured DispatchError
-    if let Some(de) = e.downcast_ref::<DispatchError>() {
-        let extra = if de.valid.is_some() || de.param.is_some() || de.hint.is_some() {
-            Some(serde_json::json!({
-                "valid": de.valid,
-                "param": de.param,
-                "hint":  de.hint,
-            }))
-        } else {
-            None
-        };
-        return (de.kind, de.message.clone(), extra);
-    }
-    // 2. ToolError serialized as JSON string (legacy radarr path)
-    let msg = e.to_string();
-    if let Ok(v) = serde_json::from_str::<Value>(&msg)
-        && let Some(kind_str) = v.get("kind").and_then(|k| k.as_str())
-    {
-        let kind: &'static str = static_kind(kind_str);
-        let message = v["message"].as_str().unwrap_or(&msg).to_string();
-        // Preserve structured extras (valid list, param name, hint) if present.
-        let has_valid = v.get("valid").is_some_and(|v| !v.is_null());
-        let has_param = v.get("param").is_some_and(|v| !v.is_null());
-        let has_hint = v.get("hint").is_some_and(|v| !v.is_null());
-        let extra = if has_valid || has_param || has_hint {
-            Some(serde_json::json!({
-                "valid": v.get("valid"),
-                "param": v.get("param"),
-                "hint":  v.get("hint"),
-            }))
-        } else {
-            None
-        };
-        return (kind, message, extra);
-    }
-    // 3. Generic fallback
-    ("internal_error", msg, None)
-}
-
-/// Map a serialized kind string to a `&'static str` from the canonical vocabulary.
-fn static_kind(s: &str) -> &'static str {
-    match s {
-        "unknown_action" => "unknown_action",
-        "unknown_subaction" => "unknown_subaction",
-        "missing_param" => "missing_param",
-        "invalid_param" => "invalid_param",
-        "unknown_instance" => "unknown_instance",
-        "auth_failed" => "auth_failed",
-        "not_found" => "not_found",
-        "rate_limited" => "rate_limited",
-        "validation_failed" => "validation_failed",
-        "network_error" => "network_error",
-        "server_error" => "server_error",
-        "decode_error" => "decode_error",
-        "confirmation_required" => "confirmation_required",
-        _ => "internal_error",
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use super::static_kind;
     use super::{Transport, resolve_port, resolve_transport};
     use crate::config::{LabConfig, McpPreferences};
-    use crate::dispatch::error::ToolError;
-
-    /// Every kind that `ToolError::kind()` can return must have an explicit arm
-    /// in `static_kind()`.  If a new variant or SDK kind is added to `ToolError`
-    /// without a matching arm here, this test will catch the silent downgrade to
-    /// `"internal_error"`.
-    #[test]
-    fn static_kind_round_trips_all_tool_error_kinds() {
-        // Fixed-variant kinds — produced by the named ToolError variants.
-        let fixed_variants: &[ToolError] = &[
-            ToolError::UnknownAction {
-                message: String::new(),
-                valid: vec![],
-                hint: None,
-            },
-            ToolError::MissingParam {
-                message: String::new(),
-                param: "p".into(),
-            },
-            ToolError::InvalidParam {
-                message: String::new(),
-                param: "p".into(),
-            },
-            ToolError::UnknownInstance {
-                message: String::new(),
-                valid: vec![],
-            },
-        ];
-
-        for err in fixed_variants {
-            let kind = err.kind();
-            assert_eq!(
-                static_kind(kind),
-                kind,
-                "static_kind({kind:?}) should round-trip but returns \"{}\"",
-                static_kind(kind),
-            );
-        }
-
-        // SDK-promoted kinds — every stable kind tag that `ApiError::kind()` can
-        // return and that `ToolError::Sdk` promotes to the top-level `kind` field.
-        // This list must stay in sync with the arms in `static_kind()`.
-        let sdk_kinds: &[&str] = &[
-            "unknown_action",
-            "unknown_subaction",
-            "missing_param",
-            "invalid_param",
-            "unknown_instance",
-            "auth_failed",
-            "not_found",
-            "rate_limited",
-            "validation_failed",
-            "network_error",
-            "server_error",
-            "decode_error",
-            "confirmation_required",
-        ];
-
-        for &sdk_kind in sdk_kinds {
-            let err = ToolError::Sdk {
-                sdk_kind: sdk_kind.to_string(),
-                message: String::new(),
-            };
-            let kind = err.kind();
-            assert_eq!(
-                static_kind(kind),
-                kind,
-                "static_kind({kind:?}) should round-trip but returns \"{}\"",
-                static_kind(kind),
-            );
-        }
-    }
 
     #[test]
     fn transport_resolution_prefers_cli_then_env_then_config() {
